Improve Task Complete Workflow
When a task completes we will now no longer bring a viewer to the foreground (unless we’re opening it for the first time). In the case that a viewer is already open, we will just get it into the correct state (respecting its focus to determine whether to fully display the new log or whether to just reload the log list so the log is available in the sidebar).

This is designed to reduce focus stealing.