Properly `metric` that returns `Value`
A metric fuction is allowed to return a `Value` (not just a `float` or `int`). Values can be lists or dictionaries, supporting a metric emitting more than a single simple value. Previous to this fix, we just cast the metric value to a `float` which is incorrect.

This commit will now properly process the metric values (including `dict` and `list`). For each entry in the `list` or `dict`, an `EvalMetric` will be produced (using the entry key or index as a suffix).