Fix scoring behaviour in basic agent (#841)
* Fix scoring behaviour in basic agentt

Inspect's basic agent will, by default, run the scorer at least twice.

It will run it at least once when checking the score, and then once for the actual scoring.

As I understand it, this is to enable retrying based whether your submission is correct. But the issue is - this means the scorer runs twice! If your scorer makes mutations to some environment state, this can cause problems with scoring correctness.

This came up in the context of an evaluation where the scorer runs some script on the disk which checks if a marker file exists, does nothing if it does, and writes this marker file at exit. This ensures the script runs once and only once on the machine. 

But, since our scorer runs twice, and the score that actually gets reported is from the second time it runs, we were seeing this script do nothing (since the marker file already existed), and report the wrong score for some reason we couldn't figure out. 

I think the fix is fairly simple - just don't run the scorer if max attempts == 1, since we will be breaking either way.

* ruff

* Update CHANGELOG.md

---------

Co-authored-by: jjallaire-aisi <joseph.allaire@dsit.gov.uk>