Fix issue with stale rendered tool input
When transition from a prism highlighted tool input to a non highlighted input (if both appear in identical positions in the overall DOM tree), the stale content would still be rendered. This is addressed by adding a key to the code element itself, which will ensure that input is remounted when the contents change (which is currently not happening since the contents are being mutated ‘behind the back’ of react by prism.