Feature/tool view improvements (#1004)
* Improve title display

If the view provides a title, use that in the tool event card and don’t display in the tool call itself.

* Properly render tool output in transcript

Since the event would provide text, that would be rendered as simple text. Instead, normalize text into a tool message, which will be rendered as text output.

* Ensure tool input uses full width

* Improve vscode dark style

When a language is present, prism can add gnarly borders to code. Remove those.

* Update CHANGELOG.md

---------

Co-authored-by: jjallaire <jj.allaire@gmail.com>