4GB + Log File Support (#1491)
* Support 64 bit central directory

If the central directory appears more than 4gb into the zip file, it can’t be properly addressed in 32 bit, so lookup the correct position.

* Support for file sizes (within the zip) that are greater than 4gb

This isn’t likely to be an issue, but for completeness, this shifts compressed and uncompressed sizes to be 64 bit, allowing for large file sizes within the zip itself.

* update changelog

---------

Co-authored-by: J.J. Allaire <jj.allaire@gmail.com>