Fix failure to restore vscode state (#2088)
* Fix failure to restore vscode state

When vscode backgrounds a tab and it is foregrounded, the viewer reloads its state from a persisted state in vscode. We filter this state for large evals, potentially excluding the complete log summary (for perf reasons). When this would happen, we would fail to reload the log file, resulting in a blank eval screen.

* Update CHANGELOG.md

---------

Co-authored-by: jjallaire <jj.allaire@gmail.com>