Preserve sample state (#2386)
The fix to screen sample list content out of the DOM caused a regression where sample state (e.g. scroll position, selection) would be list since the view was being completed unloaded and reloaded.

As a simple workaround for the time being, this moves the screening of content lower into the row rendering, allowing the list to preserve its state exactly.